[wip 4] API change: completely moved all stuff around
