+++ examples/sod/ArborX_SOD.hpp
@@ -14,45 +14,90 @@
 
 #include <ArborX.hpp>
 #include <ArborX_DetailsSOD.hpp>
-#include <ArborX_SODStruct.hpp>
 
 namespace ArborX
 {
 
-template <typename MemorySpace>
-struct SODOutputData
+namespace SOD
 {
-  template <typename T>
-  using View = Kokkos::View<T *, MemorySpace>;
-  template <typename T>
-  using BinView = Kokkos::View<T **, MemorySpace>;
 
-  View<float> sod_halo_masses;
-  View<float> sod_halo_rdeltas;
+template <typename MemorySpace>
+struct Spheres
+{
+  Kokkos::View<Point *, MemorySpace> _centers;
+  Kokkos::View<float *, MemorySpace> _radii;
+};
 
-  BinView<int> sod_halo_bin_ids;
-  BinView<int> sod_halo_bin_counts;
-  BinView<float> sod_halo_bin_masses;
-  BinView<float> sod_halo_bin_outer_radii;
-  BinView<float> sod_halo_bin_radial_velocities;
+template <typename Particles>
+struct ParticlesWrapper
+{
+  Particles _particles;
+};
+} // namespace SOD
 
-  View<int> sod_particles_offsets;
-  View<int> sod_particles_indices;
+template <typename MemorySpace>
+struct AccessTraits<SOD::Spheres<MemorySpace>, PrimitivesTag>
+{
+  using memory_space = MemorySpace;
 
-  SODOutputData()
-      : sod_halo_masses("sod_halo_masses", 0)
-      , sod_halo_rdeltas("sod_halo_rdeltas", 0)
-      , sod_halo_bin_ids("sod_halo_bin_ids", 0, 0)
-      , sod_halo_bin_counts("sod_halo_bin_counts", 0, 0)
-      , sod_halo_bin_masses("sod_halo_bin_masses", 0, 0)
-      , sod_halo_bin_outer_radii("sod_halo_bin_outer_radii", 0, 0)
-      , sod_halo_bin_radial_velocities("sod_hlo_bin_radial_velocities", 0, 0)
+  KOKKOS_FUNCTION static std::size_t
+  size(const SOD::Spheres<MemorySpace> &spheres)
   {
+    return spheres._centers.extent(0);
+  }
+  KOKKOS_FUNCTION static Box get(SOD::Spheres<MemorySpace> const &spheres,
+                                 std::size_t const i)
+  {
+    auto const &c = spheres._centers(i);
+    auto const r = spheres._radii(i);
+    return {{c[0] - r, c[1] - r, c[2] - r}, {c[0] + r, c[1] + r, c[2] + r}};
   }
 };
 
-namespace SOD
+template <typename Particles>
+struct AccessTraits<SOD::ParticlesWrapper<Particles>, PredicatesTag>
 {
+  using ParticlesAccess = AccessTraits<Particles, PrimitivesTag>;
+
+  using memory_space = typename ParticlesAccess::memory_space;
+  using Predicates = SOD::ParticlesWrapper<Particles>;
+
+  static KOKKOS_FUNCTION size_t size(Predicates const &w)
+  {
+    return ParticlesAccess::size(w._particles);
+  }
+  static KOKKOS_FUNCTION auto get(Predicates const &w, size_t i)
+  {
+    return attach(intersects(ParticlesAccess::get(w._particles, i)), (int)i);
+  }
+};
+
+namespace Details
+{
+
+template <typename MemorySpace, typename Callback>
+struct Profiles
+{
+  Kokkos::View<Point *, MemorySpace> _fof_halo_centers;
+  float _r_min;
+  Kokkos::View<float *, MemorySpace> _r_max;
+  int _num_bins;
+  Callback _callback;
+
+  template <typename Query>
+  KOKKOS_FUNCTION void operator()(Query const &query, int halo_index) const
+  {
+    int particle_index = getData(query);
+    Point const &point = getGeometry(query);
+
+    float dist = Details::distance(point, _fof_halo_centers(halo_index));
+    if (dist > _r_max(halo_index)) // false positive
+      return;
+
+    auto bin_id = binID(_r_min, _r_max(halo_index), dist, _num_bins);
+    _callback(particle_index, halo_index, bin_id);
+  }
+};
 
 template <typename MemorySpace>
 struct MassAvgRadiiCountProfiles
@@ -72,268 +117,169 @@ struct MassAvgRadiiCountProfiles
   }
 };
 
-template <typename ExecutionSpace, typename MemorySpace, typename Particles,
-          typename ParticleMasses, typename FOFHaloCenters,
-          typename FOFHaloMasses>
-void sodCore(ExecutionSpace const &exec_space, Particles &particles,
-             ParticleMasses &particle_masses, FOFHaloCenters &fof_halo_centers,
-             FOFHaloMasses &fof_halo_masses, SODOutputData<MemorySpace> &out,
-             Parameters const &params)
+} // namespace Details
+
+template <typename MemorySpace>
+struct SODHandle
 {
-  static_assert(
-      KokkosExt::is_accessible_from<MemorySpace, ExecutionSpace>::value, "");
-  static_assert(std::is_same<typename Particles::memory_space, MemorySpace>{},
-                "");
-  static_assert(
-      std::is_same<typename ParticleMasses::memory_space, MemorySpace>{}, "");
-  static_assert(
-      std::is_same<typename FOFHaloCenters::memory_space, MemorySpace>{}, "");
-  static_assert(
-      std::is_same<typename FOFHaloMasses::memory_space, MemorySpace>{}, "");
+  Kokkos::View<ArborX::Point *, MemorySpace> _fof_halo_centers;
+  float _r_min;
+  Kokkos::View<float *, MemorySpace> _r_max;
+  BVH<MemorySpace> _bvh;
 
-  auto const num_halos = fof_halo_centers.extent(0);
-  auto const num_sod_bins = params._num_sod_bins;
-
-  // Do not sort for now, so as to not allocate additional memory, which would
-  // take 8*n bytes (4 for Morton index, 4 for permutation index). This will
-  // results in running out of memory for the hardest HACC problem on V100.
-  bool const sort_predicates = false;
-
-  // Compute R_min and R_max radii for every FOF halo
-  float r_min;
-  Kokkos::View<float *, MemorySpace> r_max;
-  std::tie(r_min, r_max) = computeSODRadii(exec_space, params, fof_halo_masses);
-  Kokkos::resize(fof_halo_masses, 0); // free as not used afterwards
-
-  // Construct the search index based on spheres (FOF centers with
-  // corresponding R_max)
-  BVH<MemorySpace> bvh(exec_space,
-                       Spheres<MemorySpace>{fof_halo_centers, r_max});
-
-  SODStruct<MemorySpace> sod_struct(exec_space, fof_halo_centers, r_min, r_max);
-
-  // Compute bin outer radii
-  out.sod_halo_bin_outer_radii =
-      computeSODBinRadii(exec_space, r_min, r_max, num_sod_bins);
-
-  // Step 2: compute some profiles (mass, count, avg radius);
-  // NOTE: we will accumulate float quantities into double in order to
-  // avoid loss of precision, which will occur once we start adding small
-  // quantities to large
-  Kokkos::View<double **, MemorySpace> sod_halo_bin_masses(
-      "ArborX::SOD::sod_halo_bin_masses", num_halos, num_sod_bins);
-  Kokkos::resize(out.sod_halo_bin_counts, num_halos, num_sod_bins);
-  sod_struct.computeBinProfiles(exec_space, particles, num_sod_bins,
-                                MassAvgRadiiCountProfiles<MemorySpace>{
-                                    particles, particle_masses,
-                                    fof_halo_centers, sod_halo_bin_masses,
-                                    out.sod_halo_bin_counts});
-
-  Kokkos::resize(out.sod_halo_bin_masses, num_halos, num_sod_bins);
-  Kokkos::parallel_for(
-      "ArborX::SOD::copy_bin_masses",
-      Kokkos::RangePolicy<ExecutionSpace>(exec_space, 0, num_halos),
-      KOKKOS_LAMBDA(int halo_index) {
-        // double -> float conversion
-        for (int bin_id = 0; bin_id < num_sod_bins; ++bin_id)
-          out.sod_halo_bin_masses(halo_index, bin_id) =
-              sod_halo_bin_masses(halo_index, bin_id);
-      });
-
-  // Figure out critical bins
-  auto critical_bin_ids = computeSODCriticalBins(
-      exec_space, params, sod_halo_bin_masses, out.sod_halo_bin_counts,
-      out.sod_halo_bin_outer_radii);
-
-  Kokkos::Profiling::pushRegion("ArborX::SOD::compute_critical_bin_particles");
-
-  // Compute offsets for storing particles in critical bins
-  Kokkos::View<int *, MemorySpace> critical_bin_offsets(
-      Kokkos::view_alloc(Kokkos::WithoutInitializing,
-                         "ArborX::SOD::critical_bin_offsets"),
-      num_halos + 1);
-  Kokkos::parallel_for(
-      "compute_critical_bins",
-      Kokkos::RangePolicy<ExecutionSpace>(exec_space, 0, num_halos),
-      KOKKOS_LAMBDA(int halo_index) {
-        critical_bin_offsets(halo_index) =
-            out.sod_halo_bin_counts(halo_index, critical_bin_ids(halo_index));
-      });
-  exclusivePrefixSum(exec_space, critical_bin_offsets);
-
-  auto num_critical_bin_particles = lastElement(critical_bin_offsets);
-  printf("#particles in critical bins: %d\n", num_critical_bin_particles);
-
-  // Find particles in critical bins for each halo
-  Kokkos::View<int *, MemorySpace> critical_bin_indices(
-      Kokkos::view_alloc(Kokkos::WithoutInitializing,
-                         "ArborX::SOD::critical_bin_indices"),
-      num_critical_bin_particles);
-  Kokkos::View<float *, MemorySpace> critical_bin_distances_augmented(
-      Kokkos::view_alloc(Kokkos::WithoutInitializing,
-                         "ArborX::SOD::critical_bin_distances_augmented"),
-      num_critical_bin_particles);
+  template <typename ExecutionSpace>
+  SODHandle(ExecutionSpace const &exec_space,
+            Kokkos::View<ArborX::Point *, MemorySpace> fof_halo_centers,
+            float r_min, Kokkos::View<float *, MemorySpace> r_max)
+      : _fof_halo_centers(fof_halo_centers)
+      , _r_min(r_min)
+      , _r_max(r_max)
+      , _bvh(exec_space, SOD::Spheres<MemorySpace>{fof_halo_centers, r_max})
   {
-    auto offsets = clone(critical_bin_offsets);
-    bvh.query(
-        exec_space, ParticlesWrapper<Particles>{particles},
-        CriticalBinParticles<MemorySpace, Particles>{
-            particles, offsets, critical_bin_indices,
-            critical_bin_distances_augmented, critical_bin_ids,
-            fof_halo_centers, out.sod_halo_bin_outer_radii, r_min, r_max},
-        Experimental::TraversalPolicy().setPredicateSorting(sort_predicates));
   }
 
-  Kokkos::Profiling::popRegion();
-
+#if 0
+  template <typename ExecutionSpace>
+  void query(ExecutionSpace const &exec_space,
+             Kokkos::View<ArborX::Point *, MemorySpace> const &particles,
+             Kokkos::View<int *, MemorySpace> &offsets,
+             Kokkos::View<int *, MemorySpace> &indices) const
   {
-    // Permute found particles within a critical bin of each halo based on their
-    // distance to the center
+    bvh.query(exec_space, ParticlesWrapper<Particles>{particles},
+              SODParticles<MemorySpace, Particles>{offsets, indices});
+    // TODO: sort results based on distance
+  }
+#endif
+
+  template <typename ExecutionSpace, typename Particles>
+  void
+  computeRdelta(ExecutionSpace const &exec_space, Particles const &particles,
+                Kokkos::View<float *, MemorySpace> const &particle_masses,
+                SOD::Parameters const &params,
+                Kokkos::View<float *, MemorySpace> &sod_halo_rdeltas,
+                Kokkos::View<int *, MemorySpace> &sod_halo_rdeltas_index) const
+  {
+    // TODO: for now, this is fixed to the usual number used for profiles.
+    // But it does not have to. Need to play around with it to see what's the
+    // fastest.
+    constexpr int num_sod_bins = 20 + 1;
+
+    auto const num_halos = _fof_halo_centers.extent(0);
+
+    // Do not sort for now, so as to not allocate additional memory, which would
+    // take 8*n bytes (4 for Morton index, 4 for permutation index). This will
+    // results in running out of memory for the hardest HACC problem on V100.
+    bool const sort_predicates = false;
+
+    // Compute bin outer radii
+    auto sod_halo_bin_outer_radii =
+        Details::computeSODBinRadii(exec_space, _r_min, _r_max, num_sod_bins);
+
+    // Step 2: compute some profiles (mass, count, avg radius);
+    // NOTE: we will accumulate float quantities into double in order to
+    // avoid loss of precision, which will occur once we start adding small
+    // quantities to large
+    Kokkos::View<double **, MemorySpace> sod_halo_bin_masses(
+        "ArborX::SOD::sod_halo_bin_masses", num_halos, num_sod_bins);
+    Kokkos::View<int **, MemorySpace> sod_halo_bin_counts(
+        "ArborX::SOD::sod_halo_bin_masses", num_halos, num_sod_bins);
+    computeBinProfiles(exec_space, particles, num_sod_bins,
+                       Details::MassAvgRadiiCountProfiles<MemorySpace>{
+                           particles, particle_masses, _fof_halo_centers,
+                           sod_halo_bin_masses, sod_halo_bin_counts});
+
+    // Figure out critical bins
+    auto critical_bin_ids = Details::computeSODCriticalBins(
+        exec_space, params, sod_halo_bin_masses, sod_halo_bin_counts,
+        sod_halo_bin_outer_radii);
+
     Kokkos::Profiling::pushRegion(
-        "ArborX::SOD::permute_critical_bin_particles");
-    auto permute =
-        Details::sortObjects(exec_space, critical_bin_distances_augmented);
-    auto critical_bin_indices_clone = clone(critical_bin_indices);
-    Kokkos::parallel_for("ArborX::SOD::apply_permutation",
-                         Kokkos::RangePolicy<ExecutionSpace>(
-                             exec_space, 0, num_critical_bin_particles),
-                         KOKKOS_LAMBDA(int i) {
-                           critical_bin_indices(i) =
-                               critical_bin_indices_clone(permute(i));
-                         });
+        "ArborX::SOD::compute_critical_bin_particles");
+
+    // Compute offsets for storing particles in critical bins
+    Kokkos::View<int *, MemorySpace> critical_bin_offsets(
+        Kokkos::view_alloc(Kokkos::WithoutInitializing,
+                           "ArborX::SOD::critical_bin_offsets"),
+        num_halos + 1);
+    Kokkos::parallel_for(
+        "compute_critical_bins",
+        Kokkos::RangePolicy<ExecutionSpace>(exec_space, 0, num_halos),
+        KOKKOS_LAMBDA(int halo_index) {
+          critical_bin_offsets(halo_index) =
+              sod_halo_bin_counts(halo_index, critical_bin_ids(halo_index));
+        });
+    exclusivePrefixSum(exec_space, critical_bin_offsets);
+
+    auto num_critical_bin_particles = lastElement(critical_bin_offsets);
+    printf("#particles in critical bins: %d\n", num_critical_bin_particles);
+
+    // Find particles in critical bins for each halo
+    Kokkos::View<int *, MemorySpace> critical_bin_indices(
+        Kokkos::view_alloc(Kokkos::WithoutInitializing,
+                           "ArborX::SOD::critical_bin_indices"),
+        num_critical_bin_particles);
+    Kokkos::View<float *, MemorySpace> critical_bin_distances_augmented(
+        Kokkos::view_alloc(Kokkos::WithoutInitializing,
+                           "ArborX::SOD::critical_bin_distances_augmented"),
+        num_critical_bin_particles);
+    {
+      auto offsets = clone(critical_bin_offsets);
+      _bvh.query(
+          exec_space, SOD::ParticlesWrapper<Particles>{particles},
+          Details::CriticalBinParticles<MemorySpace, Particles>{
+              particles, offsets, critical_bin_indices,
+              critical_bin_distances_augmented, critical_bin_ids,
+              _fof_halo_centers, sod_halo_bin_outer_radii, _r_min, _r_max},
+          Experimental::TraversalPolicy().setPredicateSorting(sort_predicates));
+    }
+
     Kokkos::Profiling::popRegion();
+
+    {
+      // Permute found particles within a critical bin of each halo based on
+      // their distance to the center
+      Kokkos::Profiling::pushRegion(
+          "ArborX::SOD::permute_critical_bin_particles");
+      auto permute =
+          Details::sortObjects(exec_space, critical_bin_distances_augmented);
+      auto critical_bin_indices_clone = clone(critical_bin_indices);
+      Kokkos::parallel_for("ArborX::SOD::apply_permutation",
+                           Kokkos::RangePolicy<ExecutionSpace>(
+                               exec_space, 0, num_critical_bin_particles),
+                           KOKKOS_LAMBDA(int i) {
+                             critical_bin_indices(i) =
+                                 critical_bin_indices_clone(permute(i));
+                           });
+      Kokkos::Profiling::popRegion();
+    }
+
+    // Compute R_delta
+    Kokkos::resize(sod_halo_rdeltas_index, 0);
+    std::tie(sod_halo_rdeltas, sod_halo_rdeltas_index) =
+        Details::computeSODRdeltas(
+            exec_space, params, particle_masses, sod_halo_bin_masses,
+            sod_halo_bin_outer_radii, critical_bin_ids, critical_bin_offsets,
+            critical_bin_indices, critical_bin_distances_augmented);
+
+    Kokkos::Profiling::pushRegion("ArborX::SOD::find_sod_particles");
   }
 
-  // Compute R_delta
-  Kokkos::View<int *, MemorySpace> sod_halo_rdeltas_index(
-      "ArborX::SOD::r_delta_index", 0);
-  std::tie(out.sod_halo_rdeltas, sod_halo_rdeltas_index) = computeSODRdeltas(
-      exec_space, params, particle_masses, sod_halo_bin_masses,
-      out.sod_halo_bin_outer_radii, critical_bin_ids, critical_bin_offsets,
-      critical_bin_indices, critical_bin_distances_augmented);
-  Kokkos::resize(sod_halo_bin_masses, 0, 0); // free as not used afterwards
-  Kokkos::resize(particle_masses, 0);        // free as not used afterwards
-
-  Kokkos::Profiling::pushRegion("ArborX::SOD::find_sod_particles");
-
-  // Compute number of particles within SOD's R_delta
-  Kokkos::resize(Kokkos::WithoutInitializing, out.sod_particles_offsets,
-                 num_halos + 1);
-  Kokkos::parallel_for(
-      "ArborX::SOD::compute_sod_particles_counts",
-      Kokkos::RangePolicy<ExecutionSpace>(exec_space, 0, num_halos),
-      KOKKOS_LAMBDA(int halo_index) {
-        int count = 0;
-
-        // Sum the counts for all bins smaller than critical one
-        for (int bin_id = 0; bin_id < critical_bin_ids(halo_index); ++bin_id)
-          count += out.sod_halo_bin_counts(halo_index, bin_id);
-
-        // Add the count for the critical bin
-        count += sod_halo_rdeltas_index(halo_index);
-
-        out.sod_particles_offsets(halo_index) = count;
-      });
-  exclusivePrefixSum(exec_space, out.sod_particles_offsets);
-
-  Kokkos::resize(out.sod_particles_indices,
-                 lastElement(out.sod_particles_offsets));
-
-  // Copy critical bin particles
-  Kokkos::parallel_for(
-      "ArborX::SOD::copy_critical_bin_particles",
-      Kokkos::RangePolicy<ExecutionSpace>(exec_space, 0, num_halos),
-      KOKKOS_LAMBDA(int halo_index) {
-        int num_valid_particles_in_critical_bin =
-            sod_halo_rdeltas_index(halo_index);
-        int critical_bin_offset = out.sod_particles_offsets(halo_index + 1) -
-                                  num_valid_particles_in_critical_bin;
-
-        // TODO: linear scan, may be expensive
-        for (int i = 0; i < num_valid_particles_in_critical_bin; ++i)
-          out.sod_particles_indices(critical_bin_offset++) =
-              critical_bin_indices(critical_bin_offsets(halo_index) + i);
-      });
-  Kokkos::resize(critical_bin_offsets, 0); // free as not used afterwards
-  Kokkos::resize(critical_bin_indices, 0); // free as not used afterwards
-
+  template <typename ExecutionSpace, typename Particles, typename Callback>
+  void computeBinProfiles(ExecutionSpace const &exec_space,
+                          Particles const &particles, int num_bins,
+                          Callback const &callback) const
   {
-    auto offsets = clone(out.sod_particles_offsets);
-    bvh.query(
-        exec_space, ParticlesWrapper<Particles>{particles},
-        SODParticles<MemorySpace, Particles>{
-            particles, offsets, out.sod_particles_indices, critical_bin_ids,
-            fof_halo_centers, r_min, r_max, num_sod_bins},
+    // Do not sort for now, so as to not allocate additional memory, which would
+    // take 8*n bytes (4 for Morton index, 4 for permutation index). This will
+    // results in running out of memory for the hardest HACC problem on V100.
+    bool const sort_predicates = false;
+    _bvh.query(
+        exec_space, SOD::ParticlesWrapper<Particles>{particles},
+        Details::Profiles<MemorySpace, Callback>{_fof_halo_centers, _r_min,
+                                                 _r_max, num_bins, callback},
         Experimental::TraversalPolicy().setPredicateSorting(sort_predicates));
   }
-  Kokkos::Profiling::popRegion();
-}
-
-} // namespace SOD
-
-template <typename ExecutionSpace, typename Particles, typename ParticleMasses,
-          typename FOFHaloCenters, typename FOFHaloMasses>
-auto sod(ExecutionSpace const &exec_space, Particles particles,
-         ParticleMasses particle_masses, FOFHaloCenters fof_halo_centers,
-         FOFHaloMasses fof_halo_masses, SOD::Parameters const &params)
-{
-  Kokkos::Profiling::pushRegion("ArborX::SOD");
-
-  using MemorySpace = typename ExecutionSpace::memory_space;
-
-  SODOutputData<MemorySpace> out_device;
-
-  // Transfer data from host to device
-  Kokkos::Profiling::pushRegion("ArborX::SOD::copy_input_data_to_device");
-  Kokkos::View<ArborX::Point *, MemorySpace> particles_device =
-      Kokkos::create_mirror_view_and_copy(exec_space, particles);
-  Kokkos::View<float *, MemorySpace> particle_masses_device =
-      Kokkos::create_mirror_view_and_copy(exec_space, particle_masses);
-  Kokkos::View<ArborX::Point *, MemorySpace> fof_halo_centers_device =
-      Kokkos::create_mirror_view_and_copy(exec_space, fof_halo_centers);
-  Kokkos::View<float *, MemorySpace> fof_halo_masses_device =
-      Kokkos::create_mirror_view_and_copy(exec_space, fof_halo_masses);
-  Kokkos::Profiling::popRegion();
-
-  // Execute the kernels on the device
-  SOD::sodCore(exec_space, particles_device, particle_masses_device,
-               fof_halo_centers_device, fof_halo_masses_device, out_device,
-               params);
-
-  SODOutputData<Kokkos::HostSpace> out_host;
-
-  // Transfer data from device to host
-  Kokkos::Profiling::pushRegion("ArborX::SOD::copy_output_data_from_device");
-  auto copy_bins_to_host = [](auto &view_host, auto &view_device) {
-    auto view_mirror =
-        Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace{}, view_device);
-    Kokkos::resize(Kokkos::WithoutInitializing, view_host,
-                   view_device.extent(0), view_device.extent(1));
-    Kokkos::deep_copy(view_host, view_mirror);
-  };
-  copy_bins_to_host(out_host.sod_halo_bin_outer_radii,
-                    out_device.sod_halo_bin_outer_radii);
-  copy_bins_to_host(out_host.sod_halo_bin_masses,
-                    out_device.sod_halo_bin_masses);
-  copy_bins_to_host(out_host.sod_halo_bin_counts,
-                    out_device.sod_halo_bin_counts);
-
-  auto copy_to_host = [](auto &view_host, auto &view_device) {
-    Kokkos::resize(Kokkos::WithoutInitializing, view_host,
-                   view_device.extent(0));
-    Kokkos::deep_copy(view_host, view_device);
-  };
-  copy_to_host(out_host.sod_halo_rdeltas, out_device.sod_halo_rdeltas);
-  copy_to_host(out_host.sod_particles_offsets,
-               out_device.sod_particles_offsets);
-  copy_to_host(out_host.sod_particles_indices,
-               out_device.sod_particles_indices);
-  Kokkos::Profiling::popRegion();
-
-  Kokkos::Profiling::popRegion();
-
-  return out_host;
-}
+};
 
 } // namespace ArborX
